agentic/checkpoint-quantum-agent/requirements.txt
# Core dependencies
crewai>=1.0.0
openai>=1.3.7
pydantic>=2.5.0
requests>=2.31.0
structlog>=23.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Check Point API
urllib3>=2.1.0
certifi>=2023.11.17

# MCP Integration
mcp>=1.16.0
httpx>=0.25.2

# Kubernetes and deployment
kubernetes>=28.1.0
fastapi>=0.104.1
uvicorn>=0.24.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
pre-commit>=3.6.0

# Monitoring
prometheus-client>=0.19.0

